Add files via upload
This function calculates the distances along a great circle, theoretical delay times and back azimuth of infrasonic signals (source-to-receiver) at any given location on Earth. The function is written for a single source and a single receiver. Please refer to infra_dist_azimuth.m for arrivals at all stations of the CTBTO IMS network.